+++ libshviotqt/src/rpc/socketrpcconnection.cpp
@@ -77,7 +77,7 @@ void SocketRpcConnection::setSocket(Socket *socket)
 		emit socketConnectedChanged(true);
 	});
 	connect(socket, &Socket::stateChanged, this, [this](QAbstractSocket::SocketState state) {
-		shvDebug() << this << "Socket state changed" << (int)state;
+		shvDebug() << this << "Socket state changed" << static_cast<int>(state);
 	});
 	connect(socket, &Socket::disconnected, this, [this]() {
 		shvDebug() << this << "Socket disconnected!!!";
@@ -145,7 +145,7 @@ bool SocketRpcConnection::isOpen()
 int64_t SocketRpcConnection::writeBytes(const char *bytes, size_t length)
 {
 	//shvLogFuncFrame();
-	return socket()->write(bytes, length);
+	return socket()->write(bytes, static_cast<qint64>(length));
 }
 
 void SocketRpcConnection::writeMessageBegin()
